Full module reorganization
