+++ moncli/__init__.py
@@ -1,3 +1,2 @@
 from .entities import MondayClient
 from .constants import *
-from .enums import *
