+++ moncli/api_v1/constants.py
