+++ moncli/constants.py
@@ -1,5 +1,3 @@
-from enum import Enum
-
 DATE_FORMAT = '%Y-%m-%d'
 DATETIME_FORMAT = '%Y-%m-%dT%H:%M:%SZ'
